add support for passthru/physical devices
this sets up tgtmaps for the v0 and p0 iports and populates them based
on the LD_GET_LIST and PD_GET_LIST dcmds.

the pd probe stuff also does a PD_GET_INFO to check if the device is
a passthru disk before adding it to the tgtmap. i would have liked to
have done this in the pd_tran_tgt_init, but the tgtmap handling generates
ugly log messages if the probe fails for a target you told that you told
it about.

im mapping the logical disks and physical disks into the bus by the
identifier the controller addresses them with, not by wwns like other
modern drivers. i dont see the point in having to translate from a
property of the device to its identifer and storing a bunch of extra state
when the midlayer is already able to address things by a target index.