rework the pd code so it actually works
this makes it appear as a sas fabric instead of a legacy SPI bus. this is
necessary because PDs are addressed by both a logical id and their handle
on the bus. both of these together are too big to fit into legacy target,
so this turns them into full blown COMPLEX addresses. now the pd probe
code figures out their sas port id, and uses that as the bus identifer
for solaris to address them with. during the tgt_init, the port id is
matched to a pd_tgt structure which contains the id and handle for the
commands, and allocates a pd_lu structure to map the lun and tgt.