add the guts of asynchronous event notifications (AEN)
on mfii, AENs are basically reads of a log of interesting events. events
include the chip booting, setting the time, battery status changes,
and most useful to us, logical and physical disk changes.

each event has a monotonically increasing sequency number. you can read
previous events, and then wait for the next one to be created.

this code fetches the sequence number from the last boot of the chip,
and then fetches them up to now and waits for future changes.