mostly bring back probing physical disks when the p0 iport appears.
turns out the iports attach after mfii_attach returns, which means that
the aens can race with the creation of the tgtmaps, and particularly
wont be there by the time mfii_ld_probe was called.

this changes pd_inserted to check if the map exists before attempting
to insert the disk. pd_probe and ld_probe are run at the end of their
respective iports now.

this also means if that the iports fail to attach things wont blow up.